Vistas actualizadas

+++ sadi/departamentos/views.py
@@ -3,29 +3,35 @@ from .models import Departamento
 from .serializers import DepartamentoSerializer
 from django.contrib import messages
 from django.shortcuts import get_object_or_404
+from usuarios.permissions import IsAdmin, IsApoyo
+from usuarios.decorators import role_required
 
 # ======================CRUD=======================
 from django.shortcuts import render, redirect
 from .forms import DepartamentoForm
 
 
+@role_required("ADMIN", "APOYO")
 def gestion_departamentos(request):
     departamentos = Departamento.objects.all()
     form = DepartamentoForm()
     abrir_modal_crear = False
     abrir_modal_editar = False
+    # permisos por rol
+    puede_crear = request.user.role in ["ADMIN", "APOYO"]
+    puede_editar = request.user.role in ["ADMIN", "APOYO"]
+    puede_eliminar = request.user.role == "ADMIN"
 
     if request.method == "POST":
-        if "crear_departamento" in request.POST:
+        if "crear_departamento" in request.POST and puede_crear:
             form = DepartamentoForm(request.POST)
             if form.is_valid():
                 form.save()
                 messages.success(request, "Departamento creado correctamente.")
                 return redirect("gestion_departamentos")
-            else:
-                abrir_modal_crear = True
+            abrir_modal_crear = True
 
-        elif "editar_departamento" in request.POST:
+        elif "editar_departamento" in request.POST and puede_editar:
             departamento_id = request.POST.get("departamento_id")
             departamento = get_object_or_404(Departamento, id=departamento_id)
             form = DepartamentoForm(request.POST, instance=departamento)
@@ -33,10 +39,9 @@ def gestion_departamentos(request):
                 form.save()
                 messages.success(request, "Departamento editado correctamente.")
                 return redirect("gestion_departamentos")
-            else:
-                abrir_modal_editar = True
+            abrir_modal_editar = True
 
-        elif "eliminar_departamento" in request.POST:
+        elif "eliminar_departamento" in request.POST and puede_eliminar:
             departamento_id = request.POST.get("departamento_id")
             departamento = get_object_or_404(Departamento, id=departamento_id)
             departamento.delete()
@@ -56,7 +61,28 @@ def gestion_departamentos(request):
 
 
 # ========================API=======================
+
+from .filters import DepartamentoFilter
+from django_filters.rest_framework import DjangoFilterBackend
+
+
 class DepartamentoViewSet(viewsets.ModelViewSet):
-    queryset = Departamento.objects.all()
     serializer_class = DepartamentoSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    queryset = Departamento.objects.all()
+    filter_backends = [DjangoFilterBackend]
+    filterset_class = DepartamentoFilter
+
+    def get_permissions(self):
+        if self.request.user.role == "ADMIN":
+            return [IsAdmin()]
+        elif self.request.user.role == "APOYO":
+            return [IsApoyo()]
+        return [permissions.IsAuthenticated()]
+
+    def get_queryset(self):
+        user = self.request.user
+        queryset = super().get_queryset()
+
+        if user.role in ["ADMIN", "APOYO"]:
+            return queryset
+        return Departamento.objects.none()
